Support rngd on arm64
Although it does not do anything, as there is no CPU rng on arm64 at present.
