+++ pkg/rngd/Makefile
@@ -1,5 +1,4 @@
 IMAGE=rngd
 DEPS:=$(wildcard cmd/rngd/*.go) $(shell find cmd/rngd/vendor)
-ARCHES=x86_64
 
 include ../package.mk
