+++ pkg/rngd/cmd/rngd/rng_linux_amd64.go
@@ -37,22 +37,12 @@ import "C"
 import (
 	"errors"
 	"flag"
-	"os"
-	"unsafe"
-
-	"golang.org/x/sys/unix"
 )
 
 var disableRdrand = flag.Bool("disable-rdrand", false, "Disable use of RDRAND")
 var disableRdseed = flag.Bool("disable-rdseed", false, "Disable use of RDSEED")
 
 var hasRdrand, hasRdseed bool
-
-type randInfo struct {
-	entropyCount int
-	size         int
-	buf          uint64
-}
 
 func initRand() bool {
 	hasRdrand = C.hasrdrand() == 1 && !*disableRdrand
@@ -71,19 +61,4 @@ func rand() (uint64, error) {
 		return uint64(x), nil
 	}
 	return 0, errors.New("No randomness available")
-}
-
-func writeEntropy(random *os.File) (int, error) {
-	r, err := rand()
-	if err != nil {
-		// assume can fail occasionally
-		return 0, nil
-	}
-	const entropy = 64 // they are good random numbers, Brent
-	info := randInfo{entropy, 8, r}
-	ret, _, err := unix.Syscall(unix.SYS_IOCTL, uintptr(random.Fd()), uintptr(C.rndaddentropy), uintptr(unsafe.Pointer(&info)))
-	if ret == 0 {
-		return 8, nil
-	}
-	return 0, err
 }
