+++ pkg/rngd/cmd/rngd/rng_unsupported.go
@@ -1,4 +1,4 @@
-// +build !linux !amd64
+// +build !linux !amd64,!arm64
 
 package main
